add navbar, login page, main page, userPage, remake users and usersListPage
