+++ .idea/workspace.xml
@@ -1,7 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="eff2988d-0997-442e-873d-b1fadaaf7d40" name="Changes" comment="" />
+    <list default="true" id="eff2988d-0997-442e-873d-b1fadaaf7d40" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/src/app/components/login.jsx" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/app/components/main.jsx" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/app/components/not-found.jsx" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/app/components/userPage.jsx" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/app/components/usersListPage.jsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/App.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/App.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/api/fake.api/user.api.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/api/fake.api/user.api.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/components/navbar.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/components/navbar.jsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/components/tableBody.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/components/tableBody.jsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/components/users.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/components/users.jsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/components/usersTable.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/components/usersTable.jsx" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -31,7 +44,7 @@
     <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
     <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
     <property name="WebServerToolWindowFactoryState" value="false" />
-    <property name="last_opened_file_path" value="$PROJECT_DIR$/../react-router" />
+    <property name="last_opened_file_path" value="$PROJECT_DIR$/src/app/components" />
     <property name="node.js.detected.package.eslint" value="true" />
     <property name="node.js.detected.package.standard" value="true" />
     <property name="node.js.selected.package.eslint" value="(autodetect)" />
@@ -42,6 +55,9 @@
     <property name="settings.editor.selected.configurable" value="settings.javascript.linters.eslint" />
   </component>
   <component name="RecentsManager">
+    <key name="CopyFile.RECENT_KEYS">
+      <recent name="C:\Users\Sl\WebstormProjects\fast-company\src\app\components" />
+    </key>
     <key name="MoveFile.RECENT_KEYS">
       <recent name="C:\Users\Sl\WebstormProjects\fast-company\src\components" />
       <recent name="C:\Users\Sl\WebstormProjects\fast-company\src\api" />
@@ -88,6 +104,9 @@
       <workItem from="1642067953263" duration="1386000" />
       <workItem from="1642326351353" duration="423000" />
       <workItem from="1642331645442" duration="1950000" />
+      <workItem from="1642419675764" duration="6526000" />
+      <workItem from="1642507718209" duration="6096000" />
+      <workItem from="1642586125218" duration="1811000" />
     </task>
     <task id="LOCAL-00001" summary="first commit">
       <created>1639150314573</created>
