+++ src/app/App.js
@@ -1,14 +1,26 @@
 import React from "react";
-import Users from "./components/users";
 import Navbar from "./components/navbar";
+import { Redirect, Route, Switch } from "react-router-dom";
+import Main from "./components/main";
+import Login from "./components/login";
+import Users from "./components/users";
 
-const App = () => {
+function App() {
     return (
         <div>
             <Navbar/>
-            <Users/>
+            <Switch>
+                <Route path="/" exact component={Main}/>
+                <Route path="/login" component={Login}/>
+                {/* <Route path="/users" exact component={UsersListPage}/> */}
+                {/* <Route path="/users/:userId?" component={UserPage}/> */}
+                <Route path="/users/:userId?" component={Users}/>
 
+                {/* <Route path="/404" component={NotFound}/> */}
+                <Redirect to="/"/>
+            </Switch>
         </div>
+
     );
 };
