+++ src/app/api/fake.api/user.api.js
@@ -148,9 +148,18 @@ const fetchAll = () =>
     new Promise((resolve) => {
         window.setTimeout(function () {
             resolve(users);
-        }, 2000);
+        }, 1000);
+    });
+
+const getById = (id) =>
+    new Promise((resolve) => {
+        window.setTimeout((function () {
+            resolve(users.find((user) => user._id === id));
+
+        }, 1000));
     });
 
 export default {
-    fetchAll
+    fetchAll,
+    getById
 };
