+++ src/app/components/login.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+const Login = () => {
+    return (<h1>Login</h1>)
+}
+export default Login
