+++ src/app/components/main.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+const Main = () => {
+    return (<h1>Main</h1>);
+};
+export default Main;
