+++ src/app/components/not-found.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+const NotFound = () => {
+    return (<h1>Loading.... </h1>)
+}
+export default NotFound
