+++ src/app/components/tableBody.jsx
@@ -1,19 +1,22 @@
 import React from "react";
 import PropTypes from "prop-types";
-import _ from 'lodash'
+import _ from "lodash";
 
-const TableBody = ({data, columns}) => {
+const TableBody = ({
+    data,
+    columns
+}) => {
     const renderContent = (item, column) => {
 
         if (columns[column].component) {
-            const component = columns[column].component
-            if (typeof component === 'function') {
-                return component(item)
+            const component = columns[column].component;
+            if (typeof component === "function") {
+                return component(item);
             }
             return component;
         }
-        return _.get(item, columns[column].path)
-    }
+        return _.get(item, columns[column].path);
+    };
 
     return (<tbody>
         {data.map((item) => (
@@ -26,11 +29,11 @@ const TableBody = ({data, columns}) => {
             </tr>
         ))}
         </tbody>
-    )
-}
+    );
+};
 
 TableBody.propTypes = {
     data: PropTypes.array.isRequired,
     columns: PropTypes.object.isRequired
-}
-export default TableBody
+};
+export default TableBody;
