+++ src/app/components/usersTable.jsx
@@ -5,10 +5,16 @@ import TableBody from "./tableBody";
 import Bookmark from "./bookmark";
 import QualitiesList from "./qualitiesList";
 import Table from "./table";
+import { Link } from "react-router-dom";
+
 
 const UsersTable = ({users, onSort, selectedSort, onToggleBookMark, onDelete, ...rest}) => {
     const columns = {
-        name: {path: "name", name: "Имя"},
+        name: {path: "name", name: "Имя",
+            component: (user) => (
+                <Link to={`users/${user._id}`}>{user.name}</Link>
+            )
+        },
         qualities: {name: "Качества", component: (user) => (<QualitiesList qualities={user.qualities}/>)},
         professions: {path: "profession.name", name: "Профессия"},
         completedMeetings: {
